libminijail: Add minijail_syscall_name
This adds minijail_syscall_name to expose lookup_syscall_name so
crash reporting can convert syscall numbers to their corresponding name
to make reports more actionable.

It also changing the input parameter from int to long to match the type
used by syscall(2).

BUG=b:200326471
TEST=FEATURES=test emerge-${BOARD} chromeos-base/minijail

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform/minijail/+/4408744
Auto-Submit: Allen Webb <allenwebb@google.com>
Tested-by: Allen Webb <allenwebb@google.com>
Commit-Queue: Allen Webb <allenwebb@google.com>

+++ util.c
@@ -178,7 +178,7 @@ int lookup_syscall(const char *name, size_t *ind)
 	return -1;
 }
 
-const char *lookup_syscall_name(int nr)
+const char *lookup_syscall_name(long nr)
 {
 	const struct syscall_entry *entry = syscall_table;
 	for (; entry->name && entry->nr >= 0; ++entry)
